+++ web/app/themes/lark/config/view.php
@@ -13,9 +13,10 @@ return [
     */
 
     'paths' => [
-        get_theme_file_path('/resources/views'),
-        get_parent_theme_file_path('/resources/views'),
+        get_theme_file_path().'/resources/views',
+        get_parent_theme_file_path().'/resources/views',
     ],
+
 
     /*
     |--------------------------------------------------------------------------
@@ -28,21 +29,8 @@ return [
     |
     */
 
-    'compiled' => get_theme_file_path('/storage/framework/views'),
+    'compiled' => wp_upload_dir()['basedir'].'/cache',
 
-    /*
-    |--------------------------------------------------------------------------
-    | View Debugger
-    |--------------------------------------------------------------------------
-    |
-    | Enabling this option will display the current view name and data. Giving
-    | it a value of 'view' will only display view names. Giving it a value of
-    | 'data' will only display current data. Giving it any other truthy value
-    | will display both.
-    |
-    */
-
-    'debug' => false,
 
     /*
     |--------------------------------------------------------------------------
@@ -57,24 +45,7 @@ return [
     */
 
     'namespaces' => [
-        /*
-         | Given the below example, in your views use something like:
-         |     @include('MyPlugin::some.view.or.partial.here')
-         */
-        // 'MyPlugin' => WP_PLUGIN_DIR . '/my-plugin/resources/views',
-    ],
-
-    /*
-    |--------------------------------------------------------------------------
-    | View Directives
-    |--------------------------------------------------------------------------
-    |
-    | The namespaces where view components reside. Components can be referenced
-    | with camelCase & dot notation.
-    |
-    */
-
-    'directives' => [
-        'asset'  => Roots\Acorn\Assets\AssetDirective::class,
+        /* Given the below example, in your views use something like: @include('WC::some.view.or.partial.here') */
+        // 'WC' => WP_PLUGIN_DIR.'/woocommerce/templates/',
     ],
 ];