+++ web/app/themes/lark/resources/assets/scripts/app.js
@@ -1,9 +1,0 @@
-/**
- * External Dependencies
- */
-import 'jquery';
-import 'bootstrap';
-
-$(document).ready(() => {
-  // console.log('Hello world');
-});