+++ web/app/themes/lark/resources/views/index.blade.php
@@ -3,21 +3,16 @@
 @section('content')
   @include('partials.page-header')
 
-  @if (! have_posts())
-    <x-alert type="warning">
-      {!! __('Sorry, no results were found.', 'sage') !!}
-    </x-alert>
-
+  @if (!have_posts())
+    <div class="alert alert-warning">
+      {{ __('Sorry, no results were found.', 'sage') }}
+    </div>
     {!! get_search_form(false) !!}
   @endif
 
-  @while(have_posts()) @php(the_post())
-    @includeFirst(['partials.content-' . get_post_type(), 'partials.content'])
+  @while (have_posts()) @php the_post() @endphp
+    @include('partials.content-'.get_post_type())
   @endwhile
 
   {!! get_the_posts_navigation() !!}
-@endsection
-
-@section('sidebar')
-  @include('partials.sidebar')
 @endsection