+++ web/app/themes/lark/resources/views/partials/content-search.blade.php
@@ -1,15 +1,11 @@
-<article @php(post_class())>
+<article @php post_class() @endphp>
   <header>
-    <h2 class="entry-title">
-      <a href="{{ get_permalink() }}">
-        {!! $title !!}
-      </a>
-    </h2>
-
-    @includeWhen(get_post_type() === 'post', 'partials/entry-meta')
+    <h2 class="entry-title"><a href="{{ get_permalink() }}">{!! get_the_title() !!}</a></h2>
+    @if (get_post_type() === 'post')
+      @include('partials/entry-meta')
+    @endif
   </header>
-
   <div class="entry-summary">
-    @php(the_excerpt())
+    @php the_excerpt() @endphp
   </div>
 </article>