+++ web/app/themes/lark/resources/views/search.blade.php
@@ -3,15 +3,14 @@
 @section('content')
   @include('partials.page-header')
 
-  @if (! have_posts())
-    <x-alert type="warning">
-      {!! __('Sorry, no results were found.', 'sage') !!}
-    </x-alert>
-
+  @if (!have_posts())
+    <div class="alert alert-warning">
+      {{ __('Sorry, no results were found.', 'sage') }}
+    </div>
     {!! get_search_form(false) !!}
   @endif
 
-  @while(have_posts()) @php(the_post())
+  @while(have_posts()) @php the_post() @endphp
     @include('partials.content-search')
   @endwhile